trace2: remove warning from TryParseSettings
Remove warning about being unable to set up TRACE2 tracing from Trace2
TryParseSettings method. This method should just check to see whether
TRACE2 is enabled - if it is not, it does not need to warn the user, as we
only collect traces from those who actively choose to opt in.